[update] EXCategory 부분 로직 함수 분리  #160
